+++ assets/javascripts/kitten/components/carousel/simple-carousel.js
@@ -72,6 +72,7 @@ class SimpleCarouselBase extends Component {
 
             return (
               <div
+                key={item.key}
                 ref={node => {
                   this.elements[index] = node
                 }}