+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `SimpleCarousel` component.
 - Fix: Fix `button` and `image` position on `ContributionCard` component.
 
 ## [23.2.1] - 2018-09-04